+++ CHANGELOG.md
@@ -36,5 +36,5 @@
 
 ## 0.3.0
 - Fix problem with Bible Fact Pak™ import where existing items will be re-imported a second time
-- Move Bible Fact Pak™ import user interface into a user control
 - Refactor IQuestionImporter interface so default constructors can now be used
+- Refactor Bible Fact-Pak™ import user interface into a generic 'direct text import' user control